interface/py_losc: expose localization to py_losc.py.
1. provide python class that inherited from core.localization in python
layer code. The python class of localization is designed to make sure
the storage of matrix in np.ndarray to match with Eigen::MatrixXd,
and do some storage conversion if it needs.

2. Now, somehow, I found there will not be any error, even if the
np.ndarray storage mismatches with the Eigen::MatrixXd. This does not
make sense and make the localization class in python useless.
I should check the python binding code carefully.