curvature: change the interface of the constructor of curvature.
Change all the matrices that the curvature constructor takes from AO
based to MO based (either on CO or LO). That is grid values on AOs,
(grid_basis_value), is replaced by grid values on LOs (grid_lo).
Make such change is because of the consideration of memory cost.
1. grid_basis_value always has dimension of [npts, nbasis], which would
be very large for large systems. grid_lo has dimension of [npst, nlo],
which can be much smaller when nlo < nbasis with the setting of
localization window.
2. This change decouples the transformation of matrix from AOs to
MOs with curvature object. Such matrix transformation can be done with
other outside helper functions. In addition, this helper functions
can take the block-wise construction of AO matrices to MO matrices.
Block-wise construction is always good to save memory.