change matrix storage order to be c-style.
In liblosc, use typedef to define the internal matrix type. This makes
the change of matrix storage order very easy.

Now the storage order of internal losc matrix is switched to be row
major. This benefits the python users to use `py_losc` and `psi4_losc`
modules, since the default storage order of numpy array is row major.

If it is necessary, we change add an option to let the user choose the
matrix storage order at the building stage. May add this option later.