psi4_losc: delete LOSC class derived from psi4.core.HF class.
It turns to be that there is no need to derive class from psi4.core.HF
to have LOSC class. Since python methods are dynamically associated. We
can just update the methods in psi4.core.HF to have LOSC behaviors.
This updating is now done to `psi4.driver.scf_wavefunction_factory()`,
which is the wfn builder in psi4 for SCF procedure.